Add openidConnectClient config to test Roles
